Handle module list in a better way
Fixes a problem (segmentation fault) if another module
is added. The module load depends on null ternimation
and qeend should contain the last symbol. The compiler
was outsmarting us by removing the last symbol so the
module init crashes when .bss start all of a sudden.

Qeend now contains a regular module registration function
and initmodule checks if the current call is the last module
call. If so, it breaks out.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>